✅ : – Cover touch menu toggle with mobile e2e
