+++ frontend/src/pages/docs/md/changelog/20251101.md
@@ -107,10 +107,15 @@ What's DSPACE, you ask? You must be new around here, and if so, welcome! I'm gla
         -   [x] Adapt quest creation UI for mobile 💯
         -   [x] Test touch interactions 💯
         -   [x] Verify mobile layouts 💯
-    -   [x] Security audit 💯
-        -   [x] Review content validation 💯
-        -   [x] Check data sanitization 💯
-        -   [x] Audit authentication flow 💯
+
+-   [x] Security audit 💯
+
+    -   [x] Review content validation 💯
+    -   [x] Check data sanitization 💯
+    -   [x] Audit authentication flow 💯
+
+-   Added Playwright coverage to verify the mobile "More" menu toggle hides and reveals unpinned
+    navigation links when tapped.
 
 -   [x] **Developer Ergonomics & Infrastructure Debt** 💯 (multi‑PR initiative)
     -   [x] **Unify package management (pnpm workspaces) 💯**
